커밋!

+++ MellowTone/src/main/java/com/phoenix/main/controller/MainController.java
@@ -4,11 +4,14 @@ import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.phoenix.main.domain.MemberVO;
 import com.phoenix.main.service.GalleryService;
@@ -127,21 +130,7 @@ public class MainController {
       model.addAttribute("body","./join/join_attention.jsp");
       return "mainview";
    }
-   
-   @RequestMapping("/insert")
-	public void insert(@RequestBody MemberVO vo, HttpServletRequest request)throws Exception{
-		String post = request.getParameter("zip1") + '-' + request.getParameter("zip2");
-
-		String phone = request.getParameter("phone1") + '-' + request.getParameter("phone2") + '-'
-				+ request.getParameter("phone3");
-		String email = request.getParameter("email1") + '@' + request.getParameter("email2");
-
-		vo = new MemberVO(0, request.getParameter("id"), request.getParameter("pass1"),
-				request.getParameter("pass2"), request.getParameter("name"), post, request.getParameter("address1"),
-				request.getParameter("address2"), phone, email);
-
-		service.insert(vo);
-	}
+ 
    
    @RequestMapping("/login")
    public String login(Model model,MemberVO vo,HttpServletRequest request)throws Exception{
@@ -205,8 +194,23 @@ public class MainController {
 	   return "mainview";
    }
    
+   @RequestMapping("/deleteform")
+   public String deleteform(Model model,MemberVO vo,HttpServletRequest request)throws Exception{
+	   String id = ((MemberVO)request.getSession().getAttribute("login")).getId();
+	   String dpass = service.select(id).getPassword();
+	   
+	   model.addAttribute("title","Mypage");
+	   model.addAttribute("list",sidebar_service.select_mypage());
+	   model.addAttribute("body","./mypage/delete_check.jsp");
+	   
+	   model.addAttribute("dpass",dpass);
+	   
+	   return "mainview";
+   }
+   
    @RequestMapping("/delete")
-   public String delete(Model model,HttpSession session)throws Exception{
+   public @ResponseBody String delete(Model model,HttpSession session,String password)throws Exception{
+	   String result="FAIL";
 	   String id = ((MemberVO)session.getAttribute("login")).getId();
 	   String dpass = service.select(id).getPassword();
 	   
@@ -215,9 +219,13 @@ public class MainController {
 	   model.addAttribute("body","./mypage/delete_check.jsp");
 	   
 	   model.addAttribute("dpass",dpass);
-	   service.delete(id);
 	   
-	   return "mainview";
+	  if("1234".equals(dpass)){
+	     result="SUCCESS";
+	   service.delete(id);
+	  }
+	  
+	   return result;
    }
    
    @RequestMapping("/pocket")
